Explicitly disable sincos optimization on Android.
I uploaded this CL before, now it turned out that although it's an
issue in compiler, but it will not be solved in short term, we have
to work around in our code termporally.

We can chat in person if you want to know more details.

BUG=
TEST=trybots
Review URL: https://webrtc-codereview.appspot.com/1026006

git-svn-id: http://webrtc.googlecode.com/svn/trunk/webrtc@3358 4adac7df-926f-26a2-2b94-8c16560cd09d