Fix frame_editing_unittest.cc
The test fails since it's assuming out/testfile.yuv exists when running the test. Just opening the file at a later time than the SetUp function seems to break the test so that's not a viable solution. This CL uses a simple workaround that simply truncates the file before opening it, which works.

BUG=none
TEST=tools_unittests in Debug+Release on Mac, Win and Linux + memcheck, tsan, asan.

Review URL: https://webrtc-codereview.appspot.com/1067004

git-svn-id: http://webrtc.googlecode.com/svn/trunk/webrtc@3401 4adac7df-926f-26a2-2b94-8c16560cd09d