Add targetBitrate to VideoCodec struct.
To be used by a codec implementation. Could for instance be interpreted
as trying to fit as much as possible on one temporal layer and send
everything that doesn't fit within target bitrate on another one.

Prevents an existing hack where startBitrate is used by a codec
implementation to signify target bitrate. This hack forces a reset of
bitrate estimation to target bitrate which creates bitrate dips.

BUG=
R=mflodman@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/10479004

git-svn-id: http://webrtc.googlecode.com/svn/trunk/webrtc@5759 4adac7df-926f-26a2-2b94-8c16560cd09d