Use Java based audio as default for WebRTC.
The work landed in 4034 (use of HW AEC in AppRTC) is currently not
active by default since we build for Open SL. I missed that when I
did my initial change (since I always disabled OpenSL by GYP_DEFINES).

This CL ensures that Java based audio is used as default in WebRTC.
It would be great if we could shift over to Open SL (to cut latency)
but that would (today) mean that we can't support the HW AEC.
Hence, we are not ready to do so yet.

BUG=4034
R=kjellander@webrtc.org, perkj@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/36699004

git-svn-id: http://webrtc.googlecode.com/svn/trunk/webrtc@8040 4adac7df-926f-26a2-2b94-8c16560cd09d